update "Globally remove the Guava toolkit dependency"
