+++ hippo4j-common/src/main/java/cn/hippo4j/common/toolkit/Joiner.java
@@ -1,0 +1,90 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package cn.hippo4j.common.toolkit;
+
+import java.io.IOException;
+import java.util.Arrays;
+import java.util.Iterator;
+import java.util.Objects;
+
+/**
+ * reference google guava
+ *
+ */
+public class Joiner {
+
+    private final String separator;
+
+    private Joiner(String separator) {
+        this.separator = Objects.requireNonNull(separator);
+    }
+
+    /** Returns a joiner which automatically places {@code separator} between consecutive elements. */
+    public static Joiner on(String separator) {
+        return new Joiner(separator);
+    }
+
+    /**
+     * Returns a string containing the string representation of each of {@code parts}, using the
+     * previously configured separator between each.
+     *
+     */
+    public final String join(Object[] parts) {
+        return join(Arrays.asList(parts));
+    }
+
+    public final String join(Iterable<?> parts) {
+        return join(parts.iterator());
+    }
+
+    /**
+     * Returns a string containing the string representation of each of {@code parts}, using the
+     * previously configured separator between each.
+     *
+     */
+    public final String join(Iterator<?> parts) {
+        return appendTo(new StringBuilder(), parts).toString();
+    }
+
+    public final StringBuilder appendTo(StringBuilder builder, Iterator<?> parts) {
+        try {
+            appendTo((Appendable) builder, parts);
+        } catch (IOException impossible) {
+            throw new AssertionError(impossible);
+        }
+        return builder;
+    }
+
+    public <A extends Appendable> A appendTo(A appendable, Iterator<?> parts) throws IOException {
+        Objects.requireNonNull(appendable);
+        if (parts.hasNext()) {
+            appendable.append(toString(parts.next()));
+            while (parts.hasNext()) {
+                appendable.append(separator);
+                appendable.append(toString(parts.next()));
+            }
+        }
+        return appendable;
+    }
+
+    CharSequence toString(Object part) {
+        Objects.requireNonNull(part);
+        return (part instanceof CharSequence) ? (CharSequence) part : part.toString();
+    }
+
+}
