+++ hippo4j-message/src/main/java/cn/hippo4j/message/service/AlarmControlHandler.java
@@ -19,14 +19,15 @@ package cn.hippo4j.message.service;
 
 import cn.hippo4j.common.constant.Constants;
 import cn.hippo4j.message.dto.AlarmControlDTO;
-import cn.hutool.cache.Cache;
-import cn.hutool.cache.CacheUtil;
 import cn.hutool.core.util.IdUtil;
 import cn.hutool.core.util.StrUtil;
+import com.github.benmanes.caffeine.cache.Cache;
+import com.github.benmanes.caffeine.cache.Caffeine;
 
 import java.util.HashMap;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.TimeUnit;
 import java.util.concurrent.locks.ReentrantLock;
 
 /**
@@ -50,12 +51,12 @@ public class AlarmControlHandler {
         if (cache == null) {
             return false;
         }
-        String pkId = cache.containsKey(alarmControl.getTypeEnum().name()) ? cache.get(alarmControl.getTypeEnum().name()) : null;
+        String pkId = cache.getIfPresent(alarmControl.getTypeEnum().name());
         if (StrUtil.isBlank(pkId)) {
             ReentrantLock lock = threadPoolLock.get(threadPoolKey);
             lock.lock();
             try {
-                pkId = cache.containsKey(alarmControl.getTypeEnum().name()) ? cache.get(alarmControl.getTypeEnum().name()) : null;
+                pkId = cache.getIfPresent(alarmControl.getTypeEnum().name());
                 if (StrUtil.isBlank(pkId)) {
                     // Val meaningless.
                     cache.put(alarmControl.getTypeEnum().name(), IdUtil.simpleUUID());
@@ -77,8 +78,9 @@ public class AlarmControlHandler {
      */
     public void initCacheAndLock(String threadPoolId, String platform, Integer interval) {
         String threadPoolKey = StrUtil.builder(threadPoolId, Constants.GROUP_KEY_DELIMITER, platform).toString();
-        long milliseconds = interval * 60 * 1000;
-        Cache<String, String> cache = CacheUtil.newTimedCache(milliseconds);
+        Cache<String, String> cache = Caffeine.newBuilder()
+                .expireAfterWrite(interval, TimeUnit.MINUTES)
+                .build();
         threadPoolAlarmCache.put(threadPoolKey, cache);
         // Set the lock to prevent false sending of alarm information.
         ReentrantLock reentrantLock = new ReentrantLock();
